backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pdfplumber
rapidfuzz
python-multipart
python-dotenv
